apps/common-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-common-py"
version = "0.1.0"
description = "Agentic AI - Shared Python utilities and models"
authors = [{name = "Agentic AI Team"}]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "azure-cosmos>=4.5.0",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.14.0",
    "azure-ai-projects>=2.0.0b2",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.12.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/common"]

# [tool.black]
# line-length = 100
# target-version = ["py312"]
# 
# [tool.ruff]
# line-length = 100
# target-version = "py312"
# select = ["E", "F", "W", "I", "N", "UP", "RUF"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
markers = [
    "integration: marks tests as integration tests that make real API calls",
    "slow: marks tests as slow running"
]

